Split escalier_parser/lib.rs into multiple files
